refactor: 액세스토큰 수동재발급 로직 삭제

+++ src/main/java/subscribers/clearbunyang/domain/user/model/request/MemberSignUpRequest.java
@@ -1,6 +1,7 @@
 package subscribers.clearbunyang.domain.user.model.request;
 
 
+import java.math.BigInteger;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -20,7 +21,7 @@ public class MemberSignUpRequest {
 
     @PasswordValidation private String password;
 
-    private Integer phoneNumber;
+    private BigInteger phoneNumber;
 
     private String address;
 }
